Create stocks and stock ownerships

+++ db/migrate/20180808024408_add_balance_and_monthly_contribution_to_users.rb
@@ -1,7 +1,7 @@
 class AddBalanceAndMonthlyContributionToUsers < ActiveRecord::Migration[5.2]
   def up
-    add_column :users, :balance, :float, scale: 2
-    add_column :users, :monthly_contribution, :float, scale: 2
+    add_column :users, :balance, :decimal, precision: 7, scale: 2
+    add_column :users, :monthly_contribution, :decimal, precision: 7, scale: 2
   end
   def down
     remove_column :users, :balance
